Abstract base class and add Restore object.
